Fix `File.moveTo` implementation (#88)

+++ library/file/src/jvmMain/kotlin/io/matthewnelson/kmp/log/file/internal/-FileJvm.kt
@@ -16,15 +16,34 @@
 package io.matthewnelson.kmp.log.file.internal
 
 import io.matthewnelson.kmp.file.AccessDeniedException
+import io.matthewnelson.kmp.file.DirectoryNotEmptyException
 import io.matthewnelson.kmp.file.File
 import io.matthewnelson.kmp.file.FileNotFoundException
+import io.matthewnelson.kmp.file.FileSystemException
 import io.matthewnelson.kmp.file.IOException
+import io.matthewnelson.kmp.file.NotDirectoryException
+import io.matthewnelson.kmp.file.delete2
 import io.matthewnelson.kmp.file.exists2
-import io.matthewnelson.kmp.file.wrapIOException
 
 @Throws(IOException::class)
-internal actual fun File.moveTo(dest: File) {
+internal actual fun File.moveLogTo(dest: File) {
     try {
+        // As much as I'd prefer calling renameTo and reacting to the
+        // failure, we must check if the source File is NOT a regular
+        // file. File.renameTo can overwrite a regular file with a
+        // directory which we do NOT want to do.
+        //
+        // This non-atomic implementation should be OK here though, as
+        // moveLogTo is only utilized for log file rotation and is always
+        // called while holding a FileLock (other processes using FileLog
+        // will not be executing a rotation at the same time).
+        if (!isFile) {
+            if (!exists2()) throw FileNotFoundException(toString())
+            if (dest.isFile) throw NotDirectoryException(dest)
+
+            // dest is either NOT a regular file, or does not exist.
+        }
+
         if (renameTo(dest)) return
     } catch (t: Throwable) {
         if (t is SecurityException) {
@@ -32,10 +51,30 @@ internal actual fun File.moveTo(dest: File) {
             e.addSuppressed(t)
             throw e
         }
-        // TODO: If failure is because of destination already existing???
-        throw t.wrapIOException { "Failed to move $this >> $dest" }
+        if (t is FileNotFoundException) throw t
+        if (t is FileSystemException) throw t
+
+        val e = FileSystemException(this, dest, "renameTo failure")
+        e.addSuppressed(t)
+        throw e
     }
 
-    if (!exists2()) throw FileNotFoundException("$this")
-    throw IOException("Failed to move $this >> $dest")
+    // At this point, source File is a regular file and exists, otherwise
+    // a FileNotFoundException would have already been thrown. Destination
+    // must be a directory, or we're on Windows and it exists. Either way,
+    // try to delete it and retry File.renameTo. Alternatively, both source
+    // and dest are directories and dest is non-empty in which case, this
+    // will fail with DirectoryNotEmptyException as intended.
+    val t = try {
+        dest.delete2(ignoreReadOnly = true, mustExist = false)
+        if (renameTo(dest)) return
+        null
+    } catch (t: Throwable) {
+        if (t is DirectoryNotEmptyException) throw t
+        t
+    }
+
+    val e = FileSystemException(this, dest, "renameTo failure")
+    t?.let { e.addSuppressed(t) }
+    throw e
 }
